Rename `keyset` module to `config`.
