+++ timeboost-utils/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod keyset;
+pub mod config;
 pub mod load_generation;
 pub mod types;
 pub mod until;
